+++ handlers/product-move-api/cfn.yaml
@@ -129,26 +129,6 @@ Resources:
           App: product-move-api
           Stack: !Ref Stack
           Stage: !Ref Stage
-          invoicingApiUrl:
-            !Sub
-            - '{{resolve:secretsmanager:${Stage}/InvoicingApi:SecretString:InvoicingApiUrl}}'
-            - { Stage: !Ref Stage }
-          invoicingApiKey:
-            !Sub
-            - '{{resolve:secretsmanager:${Stage}/InvoicingApi:SecretString:InvoicingApiKey}}'
-            - { Stage: !Ref Stage }
-          zuoraBaseUrl:
-            !Sub
-            - '{{resolve:secretsmanager:${Stage}/Zuora/User/ZuoraApiUser:SecretString:baseUrl}}'
-            - { Stage: !Ref Stage }
-          zuoraUsername:
-            !Sub
-            - '{{resolve:secretsmanager:${Stage}/Zuora/User/ZuoraApiUser:SecretString:username}}'
-            - { Stage: !Ref Stage }
-          zuoraPassword:
-            !Sub
-            - '{{resolve:secretsmanager:${Stage}/Zuora/User/ZuoraApiUser:SecretString:password}}'
-            - { Stage: !Ref Stage }
       CodeUri:
         Bucket: !Ref DeployBucket
         Key: !Sub ${Stack}/${Stage}/product-move-api/product-move-api.jar
