+++ handlers/product-move-api/src/main/scala/com/gu/productmove/refund/RefundHandler.scala
@@ -3,7 +3,8 @@ package com.gu.productmove.refund
 import com.amazonaws.services.lambda.runtime.events.SQSEvent
 import com.amazonaws.services.lambda.runtime.events.SQSEvent.SQSMessage
 import com.amazonaws.services.lambda.runtime.{Context, RequestHandler}
-import com.gu.productmove.invoicingapi.{InvoicingApiRefundLive, Secrets}
+import com.gu.productmove.Secrets
+import com.gu.productmove.invoicingapi.InvoicingApiRefundLive
 import com.gu.productmove.refund.*
 import com.gu.productmove.zuora.{
   CreditBalanceAdjustmentLive,
