+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/RefundSupporterPlusSpec.scala
@@ -6,7 +6,8 @@ import com.gu.productmove.{AwsCredentialsLive, AwsS3Live, GuStageLive, SQSLive, 
 import com.gu.productmove.endpoint.available.{Billing, Currency, MoveToProduct, Offer, TimePeriod, TimeUnit, Trial}
 import com.gu.productmove.endpoint.move.ProductMoveEndpoint
 import com.gu.productmove.endpoint.move.ProductMoveEndpointTypes.ExpectedInput
-import com.gu.productmove.invoicingapi.{InvoicingApiRefundLive, Secrets}
+import com.gu.productmove.invoicingapi.InvoicingApiRefundLive
+import com.gu.productmove.Secrets
 import com.gu.productmove.zuora.GetSubscription
 import com.gu.productmove.zuora.Subscribe.*
 import com.gu.productmove.zuora.GetSubscription.GetSubscriptionResponse
